refactor : #397 Review 경로 변경
